requirements.txt
lmdb==0.97
numpy==1.17.2
Pillow==9.0.0
six==1.12.0
torch==1.2.0
torchvision==0.4.0
opencv-python